Fix custom resources being confused for game resources on F10 reload
Occasionally after an F10 reload we see some custom resources being
confused for game resources, which in certain circumstances may cause a
checktextureoverride to execute the wrong command list, texture
filtering to return an incorrect value, and potentially other subtle
issues that might be chalked up to "glitches".

Such a sequence was observed where a texture used in two subsequent draw
calls was replaced, then the replaced texture was mistaken for a
completely unrelated texture that also had a replace rule, causing the
texture to turn purple depending on the draw order in the frame:

    000211 PSSetShaderResources(StartSlot:0, NumViews:1, ppShaderResourceViews:0x00000076B88FF760)
           0: view=0x00000208B9D9B2E8 resource=0x00000208B9F262A0 hash=33502c6a
    ...
    000211 DrawIndexed(IndexCount:138, StartIndexLocation:33252, BaseVertexLocation:0)
    ...
    000211 3DMigoto   [commandlist\mods\textures\d3dx.ini\textures] checktextureoverride = ps-t0
    000211 3DMigoto    [TextureOverride\Mods\Textures\d3dx.ini\Mary] ps-t0 = resourcemary unless_null
    ...
    000212 3DMigoto   [commandlist\mods\textures\d3dx.ini\textures] checktextureoverride = ps-t0
    000212 3DMigoto    [TextureOverride\Mods\Costumes\White\tan.ini\White] ps-t0 = resource\mods\tan\tan.ini\tan1 unless_null

Additional similar glitches were also observed, where a mesh was
replaced by a completely unrelated mesh.

I believe this is a similar issue to the issue with shaders being mixed
up that was fixed by e7cb58fca6a1be3d327567744d3a6335447a558e. In this
case a resource that we were replacing was released by the game, and
later 3DMigoto happened to create a custom resource sharing that same
address. When that resource was bound, it was mistaken for the earlier
resource that had been freed.

This takes a slightly different approach to fix this than the above
shader issue, and this is somewhat of an experiment to see how viable
this approach is. This uses the fact that SetPrivateDataInterface() will
call Release() on the attached interface when the resource is freed,
providing us with a way to track the resource lifetime without actually
wrapping or hooking the resource itself, so we can purge it from our
global data structures.

This is an experiment because we know that the performance of
Get/SetPrivateData is bad, and must be avoided in hot paths. But since
we only need to call those APIs once when the resource is created the
hope is that will be negligeable in practice, and this is a lot less
invasive than the shader fix was, and less likely to be broken again in
the future. If this goes well, we might consider switching the shader
fix to do the same.