Resource Release Tracking: micro-optimize
This eliminates an unnecessary Release call from the constructor time a
resource is created and removes the helper function.

This is unlikely to make any significant difference, but atomic does not
mean fast, and these extra calls were pretty pointless - the
SetPrivateDataInterface call will take care of bumping the refcount, so
everything will still balance, and (other than running out of memory)
there shouldn't be any circumstance in which that call would fail to
necessitate the helper.

In fact, based on the debug logging it appears the refcount is never
bumped on the tracker objects after they are associated at all (i.e.
it's not bumped every time the resource's refcount is bumped), and we
could probably quite safely drop the atomic or even eliminate the
refcounting altogether and just treat the Release() call as final, but
that would be breaking away from COM.