Upscaling: Make "bb" work regardless of custom shader order
When upscaling is in use, other shaders rendering to the overlay must
use "f_bb" if they are run before the upscaling shader, and "bb" if they
are run after it. This makes it inconvenient to switch the order of
these shaders around, and can easily lead to issues where an overlay
will work for the shaderhacker, but won't work for a user that has
turned on upscaling.

To solve this, we add a new "special = upscaling_flip_bb" command that
changes the meaning of "bb" mid-frame, so that if it is used before
upscaling, it will refer to the fake back buffer, and if it is used
afterwards it will refer to the real back buffer. For backwards
compatibility, if this command is not found anywhere in the d3dx.ini,
"bb" will continue referring to only the real back buffer.

We also add a "r_bb" target that always refers to the real back buffer,
that the upscaling shader itself can use, which is particularly
important if it were to use used in a 3D Vision Direct game where
upscaling needs to be run twice.