Command List: Add <iniparam> = scissor_<side> to get scissor rectangle
Ordinarily the scissor clipping rectangle(s) cannot be stereoised, and
we have traditionally taken the approach of just disabling scissor
clipping. However, in some games scissor clipping is necessary to
prevent certain effects from being broken, but we may still want to be
able to adjust the depth of certain things that have scissor clipping
applied - the HUD being the most common.

This adds a way to pass the scissor rectangle into the shader via
IniParams, so that the shader may implement its own form of scissor
clipping. e.g. the pixel shader may compare SV_Position with a stereo
corrected version of the scissor rectangle and issue the "discard"
instruction if it falls outside of the rectangle.