Hooking: Fix incorrect use of GetImmediateContext to get HackerContext
There are a couple of places where we assume the result of
HackerDevice->GetImmediateContext() will be a HackerContext and cast it
directly, but that is not the case when hooking is enabled, as it will
just return the hooked DirectX object instead, which we cannot cast,
causing possible crashes at swap chain creation and definite crashes at
config reload.

This fixes the former by just using GetImmediateContext to create the
HackerContext if required and bump the refcount, but still accessing it
via GetHackerContext. The later is not clear we ever needed
GetImmediateContext, and if we do then the [Present] command list will
also be broken, so just change it to use GetHackerDevice and issue a
dire BUG warning if that doesn't work.