Texture Filtering: Change how filter_index is used with multiple matches
If multiple TextureOverride sections match a given texture filter, we
want to use the highest priority section to determine the filter index,
but we don't check if that section has actually set a filter index,
leading to certain cases the user may not expect:

    [ShaderOverride Uses Texture Filtering]
    hash = ...
    x = ps-t0

    [TextureOverride AAA Intended For Use With Texture Filtering]
    hash = 12345678
    filter_index = 2

    ...

    [TextureOverride ZZZ Intended For Use With Mesh Replacement]
    hash = 12345678
    match_first_index = 0
    vb0 = ResourceCustomVB
    ib = ResourceCustomIB
    drawindexed = auto
    handling = skip

In this case, since neither section indicated a match_priority their
priority will be determined by the section name, so the mesh replacement
section will take priority over the texture filtering section, and the
texture filter will be set to 1 ("some" section matched) instead of the
2 (specific section matched) the user probably expected.

This patch changes the behaviour so that the highest priority section
*that set a filter_index* will be used. This does have the potential to
change the behaviour of existing fixes, but since it is only relatively
recently that multiple TextureOverride sections could even match the
same hash, the impact should be minimal, and mostly for the better.

This also helps a little in an environment with multiple mods from
different authors, as it helps limit texture filtering to sections
actually intended for texture filtering. This isn't a complete solution
for that environment though - ideally we would have a way to restrict
the match to a subset of TextureOverrides appearing in the same
namespace. But narrowing down potential TextureOverrides is something we
may want even outside of that environment, so improvements there can
come later.