Fix ignoring directories with other attributes while searching for config files
If a recursively included directory has some other attribute set, such
as may happen when it is extracted from a 7z file, 3DMigoto would ignore
it while searching for config files to include. Fix this, and add more
logging to the include processing to make it clearer what 3DMigoto is
doing to identify issues.