Add next_marking_mode key & display in overlay
This adds a key that can cycle through available marking modes rather
than having to edit the d3dx.ini and reload each time. The current
marking mode is displayed in the overlay if it is not the default skip
mode.

Since the marking mode can now be changed by key input, it is nice to
avoid resetting it every time the config is reloaded. We still want to
reset if it is actually changed though so existing workflows continue to
work, so remember the previous setting and compare it on config reload.

Zero marking mode is skipped, since that has known crashes.

Requested-By: DJ-RK