Change default hunting key bindings using numpad dot
This uses the numpad dot key as a modifier for hunting key bindings, so
that we can add bindings for compute, hull, domain and geometry shaders
on the same keys as used for cycling through vertex and pixel shaders,
render targets and index buffers with the modifier held. It changes
next_marking_mode key to numpad dot + numpad 0.