Sort presets for a consistent order
If multiple presets activate in the same frame and touch the same
values, the order in which they run, and therefore which one actually
takes effect will be inconsistent and unpredictable.

This is one case where we need to use a sorted map to make sure the
order is always consistent. It might be nice to also add an explicit
match_priority, similar to the fuzzy TextureOverrides, but we would need
to use a different data structure to achieve that.

Requested-By: DJ-RK