Command List Profiling: Give ShaderRegex matches a name
If a ShaderRegex matches a shader and has a command list, it can create
a new ShaderOverride data structure for that shader to link the command
lists together, but since the command lists in the new ShaderOverride
did not go through the regular parsing routines they lack their ini
section for the profiling dump. Have these copy the ShaderRegex ini
section in, with .Match=<hash> on the end to differentiate them from the
list they link to.