Resource Copying: Don't strip first vertex/index from buffers
To try to simplify things when copying a vertex or index buffer 3DMigoto
attempted to strip off any part of the buffer prior to the bound offset
and first index/vertex of the current draw call.

This however created a problem where this would not work correctly if
the draw call had a non-zero first index:

    [ResourceBakIB]

    [TextureOverrideFoo]
    hash = ....
    ResourceBakIB = ref ib
    ib = ref ResourceBakIB

The problem here is that we handle cutting off "first" by treating it as
an additional "offset", but when we later rebind the buffer we use a
modified offset and have not restored the original buffer that we
intended.

We don't want to do this cut off at all when copying an index/vertex
buffer by reference, and it is debatable whether we should even really
do it when copying by value. We do need to cut off the "offset" part
when copying by value, since if we bind it to a view there is no offset
field, but the first index and index count should be handled separately.

For now this removes the "first" vertex/index cutoff so that rebinding
the index buffer will work, and if necessary we can re-visit cutting it
off when binding the buffer as a shader resource view at a later date,
or alternatively we could pass it to the shaders via ini params.