Profiling: Move UpdateProfilingTxt into Overlay
The profiling text is only used by the overlay at present, so move it
into the overlay source file to update it when it is needed.

This also means that this particular source of overhead will be measured
as part of the overlay overhead rather than the present call overhead,
so that we can focus on the real issues with the present call's
performance.