Profiling: Make profiling text safe from divide by zero errors
If something we are dividing by in the profiling text happens to be zero
(e.g. by starting profiling on the same frame we are rendering the text)
we don't want to risk crashing, so skip anything that may divide by
these when they are zero, and bail if the performance counter frequency
is reported as 0Hz for whatever reason.

This means the profiling data will be displayed one frame after enabling
profiling for instant feedback, then will continue profiling and
updating once per second as usual.