Profiling: Prevent present overhead starting very large
When profiling is first enabled the present overhead will briefly
display as a very large number before correcting itself on the next
interval.

This is due to the fact that profiling gets enabled in the middle of the
region of code that is profiled for the present call, but since it
wasn't enabled the starting time for this region of code was not updated
and the overhead for that frame will be vastly exaggerated.