Command List Profiling: Give ShaderRegex link commands an ini line
The ShaderRegex sections can synthesize a command to run a linked
command list, which will show up as a blank line in the top commands.

Similar to the synthesized command list, fill in the ini line with the
ShaderOverride section (or synthesized ShaderRegex.Match section) and a
synthesized "run" command linked to the main ShaderRegex section.