Profiling: Profile 3DMigoto's GPU overhead
This performs asynchronous timestamp queries on the GPU to determine
what impact 3DMigoto is having on the GPU and reports them on the
profiling summary page. It profiles GPU usage in the present call, draw
calls, map/unmap and overlay. The command list GPU usage is not yet
profiled.

However, this profiling has itself caused a major performance issue and
absolutely destroys the framerate. There are some known issues, such as
blocking when updating the overlay until the GPU queries have all been
retrieved, but the framerate issue it causes is besides that.

Exactly what is causing this performance issue will need to be
investigated if we are to ever include this in 3DMigoto for real. Some
theories:

- This will be performing a lot of timestamp queries, and perhaps the
  GPU is simply not optimised to be able to perform them very quickly.

- Perhaps the fact that the disjoint query only works occasionally might
  provide some clue that there is something more fundamentally wrong.

- Perhaps the increased traffic in the GPU command stream has hit a
  performance wall.

- Perhaps allocating the queries each time is causing some overhead in
  the VRAM management that could be alleviated by using a query pool.

- Perhaps the C++ data structures we are using to store the query
  references are performing excessive work.

If this turns out to be unworkable, we might consider other approaches
in the future, such as a smaller randomised sample each frame that is
extrapolated to estimate the overall performance.