Profiling: Reduce performance hit introduced by GPU profiling itself
This helps alleviate the performance issues associated with measuring
the GPU performance by implementing a query pool, so that instead of
creating new queries every time we want the timestamp on the GPU we can
reuse one we created previously. In Firewatch this returns the framerate
back to 60fps, but by itself still stalls every time the overlay text is
updated.

To alleviate the stalling this tallies up all the GPU timestamp queries
every frame and returns them to the pool to avoid having to do them all
at once when we are ready to update the timestamp text. Once we are
ready to update it, we put a hold on adding any new queries while we
wait for the disjoint query to complete, and do not block while that
query is in progress.

In Firewatch this restores the framerate to 60-70fps with profiling
enabled and no stuttering, but in DOAXVV we still see only 25fps - an
improvement from the 15fps we were hitting before, but nowhere near the
60fps we were getting when only measuring CPU performance. If this is
the performance hit we are taking before we have even measured the
command list performance this seems unworkable, and we may have to
resort to something like randomised sampling if we want to measure GPU
performance in the future.

Somewhat suspect, is that we see the map/unmap GPU overhead being
measured fairly large, with something like 1ms/frame and reportedly
costing 1-2fps, but 3DMigoto shouldn't be adding any extra GPU calls
here. This may be down to just the shear number of Map/Unmap calls
adding up since the timestamps will still increase, but it still seems
suspect.