Command Lists: Optimise out noop commands
Some commands may effectively be noops, which will cost a small amount
of processing time each frame and this could add up if there are many of
these. This may include commands that are implicitly added to both pre
and post lists but only the pre command actually does any work, or
commands that only do something if hunting or frame analysis is enabled.

This adds noop checks to certain commands that could be noops, and
optimises these out after all the command lists have been parsed. This
process is repeated while at least one command is found to be a noop, so
that recursive commands such as "run" and "checktextureoverride" can be
optimised out if all the commands they could run are also noops.

checktextureoverride is treated specially, as it could potentially run
any of the TextureOverride command lists, and as such can only be
treated as a noop if every TextureOverride command list (of matching pre
or post) is empty. This is a candidate for further optimisation if we
can add additional constraints on TextureOverride sections, such as
which slots they are allowed to match, or which checktextureoverride
commands they apply to.

Most of the command lists are registered for later optimisation as they
are parsed, but the non-fuzzy TextureOverride sections have to be
deferred, as the pointers to their command lists become invalidated as
the TextureOverride data structures (specifically the vector of
TextureOverrides sharing the same hash) are manipulated.