Profiling: Only enable command list profiling in relevant displays
This disables profiling the command list and their commands when an
unrelated profiling display is shown.

The command lists are still profiled when the summary page is shown,
since they are tallied up as part of that display, but the individual
commands do not need to be profiled for that display.