Hash Tracking: Add option to disable contamination detection
Even when full hash tracking is disabled, we still enable some minimal
hash contamination detection when in hunting mode to help alert the
shaderhacker when they might need to enable full hash tracking.

However, this contamination detection does itself have a performance
impact, and while it is limited to hunting mode, it is still nice to be
able to recover this performance once the shaderhacker is certain that
they will not be using this option, so add an =2 mode to completely
disable both hash tracking and contamination detection.

Regression tested in Far Cry 4, where track_texture_updates is
necessary, and the performance monitor estimates it can consume between
2-10fps depending on the situation and roughly corresponding to some
stuttering seen in that game.