Command Lists: Add if/else/endif flow control
These commands add the long-promised conditional logic to 3DMigoto,
which vastly expands the flexibility of the command lists. As well as
providing additional functionality, this can be particularly important
for performance in some cases since we can now choose to not do work
when we don't need to, or choose between several options (e.g. for
Single GPU vs SLI).

The if/else/endif syntax was chosen over the HLSL/C style "if () {} else
{}" because the later is reminiscent of programming languages where
newlines are optional, and that is not the case here. Python style is
out because our use of the ini format means we strip whitespace and
cannot use indentation for anything meaningful. if/else/endif seemed a
good choice since it is used in languages where the newline is
significant, and the similarity to shader assembly is a bonus. The
"then" found in BASIC languages was not used, because it is pointless.

To keep this syntax clean, the ini parser is relaxed to allow command
list sections to have lines without an equals sign.

Presently the if command only take a single operand and can only test it
for zero/non-zero, but this can be expanded to a complete expression
parser in the future.

The parsing works by inserting "if" and "else" commands as usual, and
then when the "endif" command is parsed it will find the most recent
"if" and "else" commands, and chop up the command list to place the
commands inside command lists in the "if" command, discarding the "else"
command placeholder. They dynamically allocated command lists are given
synthesized ini section names based on the if/else section and line.

There is no "else if" or "elif" command as yet to keep things
(relatively) simple, since that command is redundant and can be
implemented with recursive if commands. This may be added later.

The command lists inside the "if" command are dynamically allocated,
since the if command could itself be optimised out and those command
lists freed, resulting in invalid pointers in the registered command
lists vector still being used for optimisation. These command lists are
managed in a second vector using shared_ptrs to ensure they remain valid
until the optimisation stage has completed.

The if command will attempt to statically evaluate its expression when
it is parsed, allowing for the entire if block to be optimised out if
the expression statically evaluates to zero/false. This is useful to use
with "if stereo_active", or (coming soon) "if sli" to entirely eliminate
the performance hit associated with commands that do not apply when
those conditions are not met.