Keys/Presets: Turn "condition" option into an operand
This changes the "condition" option in Key and Preset sections to use
the same operand processing as the command list, effectively expanding
the available options to use as a condition, and later once we have an
expression parser this will allow the condition to be more than just a
single operand evaluating to zero/non-zero.

None of the operands that require a CommandListState to evaluate can be
used in this context, though we could relax that in the future as most
of those operands only use the state object for cache... though it's not
terribly likely that we would actually want to use any of these in this
context anyway.