Overlay: Add monospaced log level
This adds a new log level for messages that should be monospaced, such
as to show syntax errors from the expression parser. This adds a new
field to the log_level structure with a pointer to one of the font
members of the Overlay class, allowing it to be specified per log level.