Expression Parser: Implement second operator and misc improvements
This implements the inequality operator, and makes a few more
improvements to the infrastructure. The factory pattern is used to
contruct operators now, so that a list of operators with the same
precedence can be passed into the transform_operators() function so that
it can match them all left to right.

Various other cleanups and fixes are made to the class hierarchy and
comments are added as this code is becoming closer to final.