Expression Parser: Implement walker helper and further cleanups
This uses a CommandListWalkable base class to signify which nodes may
contain recursive syntax trees that can be walked over, and returns a
vector of those syntax trees.

This also cleans up a lot of the excessive logging, switching some to
debug logging and removing others. The debug log will now only show the
tokenisation and final syntax trees, not the intermediate stages.