Expression Parsing: Fix recursive walking not visiting all nodes
Certain expressions like "if 1 + (5 >= 0.09)" were giving errors as the
>= operator had not been transformed. The way walking was implemented
only allowed it to visit child syntax trees directly in the current
level, which would exclude the syntax tree grouped on the RHS of the
addition operator, since there was an operator between the root syntax
tree and the child one.

Change the Walkable classes to walk all walkable classes rather than
just syntax trees, and split the transform routine up into a recursion
routine that walks all walkable nodes depth first, and a visitor routine
that only visits syntax trees. This allows operators to be transformed,
no matter how they are nested.