Expression Parsing: Implement right-associative unary operators
This implements the unary +, - and ! operators. For now this uses the
same base class as the binary operators, but the LHS will be set to
nullptr and evaluated as NaN (in case it gets inadvertently used).

The operators will only parse if there is not an operand to their left,
which is important so they do not trump the regular binary addition and
subtraction operators that have lower precedence.