Expression Parsing: Implement exponentation operator, fix logical and/or
This implements the right-associative binary ** operator for computing
powers and roots. It's precedence is set just below unary operators to
make sure that 4**-2 works as expected for the square root.

This also splits logical and and or into two separate precedence groups
with and taking higher precedence than or, which matches the behaviour
of other languages like C, HLSL and Python.