Clear Command: Fix channel propagation not working
When the clear command is used with limited channels, e.g.

    clear = bb 1

It is supposed to propagate the value from the last specified channel to
all remaining channels. i.e. the above is supposed to clear the back
buffer with white. However, due to an extra increment it would propagate
the next channel instead, so the above would clear the back buffer with
red.