3DVision2SBS: Implement SLI Optimised Version
stereo2mono can present a bottleneck at high resolutions in SLI,
severely harming the framerate. To alleviate this performance hit we can
downscale the buffer prior to using stereo2mono to reduce the amount of
data that has to be copied between GPUs. The 3DVision2SBS shader is a
good candidate for this optimisation since it operates at half
resolution anyway, so if we downscale before the stereo2mono copy it
will appear the same as if we downscaled when copying the result to the
back buffer.

However, since the SBS shader can operate in either half-width or
half-height, until now we could not automatically select the correct
downscaling mode to use, and on a single GPU system running the extra
shader for downscaling would have a slight performance hit in itself.

Now that we have conditional logic we can solve this, by selecting
the correct downscaled buffer for the output mode, and running the
downscale shader only when necessary. We can even skip the entire cost
of running the shader when 3D Vision is disabled or the output mode is
set to normal.

This also fixes a couple of small off by half a pixel accuracy errors in
the regular 3DVision2SBS shader for single GPU users.