Expression Parser: Add === and !== operators for identical tests
The equality operator falls short in some cases, such as for testing for
negative zero, which is always considered to be equal to positive zero.

However, we do use negative zero in texture filtering to signify that no
texture is bound in a given slot, and this may also be useful to test
from the command list. The === and !== operators test if both operands
are binary identical, allowing for negative zero to be tested.