Command List: Change all BUG DoubleBeepExits into dire warnings
If we hit any of these BUG cases something is wrong and we need to know
about it to be able to fix it, but we sometimes just get reports that
3DMigoto "crashed, and oh yeah it beeped a little", which doesn't tell
us which BUG case was triggered. Change the strategy we use for these to
display a dire warning on the overlay, but try to continue anyway.
Hopefully then the bug reports will include an indication of which
triggered without having to go back and request more info.

This change was prompted by one of these triggering when 3DMigoto is
built with the vs2013 toolchain that doesn't trigger when built with the
vs2017 toolchain.