Expression Parsing: Work around VS2013 Bug/limitation
We use constructs like scanf("%f%n", &val, &pos) throughout 3DMigoto to
get the number of characters that scanf parsed for various reasons, such
as warning if only part of a string was parsed.

In the expression parser, we already had a pos that was a size_t, not
the int that %n expects. The vs2017 toolchain indicated that the %zn was
required for this case, and that worked fine there. However, back on the
vs2013 toolchain this does not work, and the pos field is left as 1 from
the earlier resource slot parsing attempt.

Since the expression parser used this to find the end of the float and
delimiter to the next token, it would treat "0.7" as a float "0" with
".7" left over. It would then find that scanf could parse ".7" as a
float as well, however this time pos was left as 0, so it passed "" to
the operand parser, which failed and triggered our BUG warning.

So the takeaway is to always use "int" with "%n" and not use any size
modifiers. In this case to make it explicit I use a separate ipos
variable of type int, and add a comment when that is set to pos. I could
have changed pos to an int and cast it wherever it was used as a size_t,
but this was is more explicit that there is a toolchain bug/limitation
that we are working around.