Infer situations where raw buffer views are required
If we have a buffer that allows raw views it may be accessed raw, or
not. However, if the full data type of the buffer is not known we will
not be able to access it as a typed buffer, so in that case we can
pretty safely infer that it must be accessed raw, allowing us to skip
the raw keyword in that case.

Note that we also don't know the type of structured buffers, however the
structured and raw access flags are mutually exclusive on the underlying
resource, so we can assume that any buffer that allows raw access will
not be structured.