Add vertex buffer hunting
This works in much the same way as index buffer hunting, and should
significantly speed up the use cases for modding communities that are
replacing meshes as vertex buffers tend to be more precise than index
buffers for the purpose of matching a mesh.

By default this takes the /*- keys on the numpad, which were previously
used for render target hunting. Render target hunting is almost never
actually helpful to identify a render target (frame analysis is much
superior for this purpose), so these keys are assigned to
home/insert/pgup and commented out.

In addition to the requirements for hunting, this also hooks up the stat
collection for vertex buffers that is reported in the log whenever a
vertex buffer, vertex shader or pixel shader is marked.