Copy index/vertex buffer hash to clipboard on mark
When index & vertex buffers are marked we log their hash and what
shaders they are used with, which is not a terribly good workflow for
people working with these hashes a lot. This copies their hash to the
clipboard so that they can easily be pasted in a TextureOverride or
elsewhere.