Remove take_screenshot dependency on hunting=1/2
It seems like a bit of an odd choice making take_screenshot dependent on
hunting mode, as people may wish to use it as an alternative to Alt+F1
even in release mode. Remove the dependency to allow this, but still
leave it in the Hunting section for historical/compatibility reasons.