Add marking_actions to select HLSL/clipboard actions
There are more and more use cases where we want to locate a shader
without decompiling it - e.g. we may just want it's hash to add to a
configuration file, or we might only want to work with assembly shaders.

This adds an option that allows the shaderhacker to choose what happens
when a shader is marked - it adds has two options that can be used
either separately or together to decompile the shader to HLSL and/or to
copy it's hash to the clipboard. Later, we can look at adding another
action to disassemble the shaders, or even try to automatically guess
which should be used.