Add snapshot_if_pink option to marking_actions
The snapshots on mark are generally not very useful to while using skip
marking mode since they show everything except what we want to know. On
the other hand they are immensely useful when using pink marking mode
where the selected object/shader is highlighted.

This adds an option to the marking actions so that the snapshots will
only dump out when in pink marking mode, to try to take advantage of
them when they will be useful, but skip them when they will likely not
be. The option can be disabled to have them dump out regardless of
marking mode as before.