Merge mark_snapshot into marking_actions
Creating a mono or stereo snapshot on mark are really just two possible
actions we might do when marking, so merge the option into
marking_actions. This also allows snapshots to be taken when marking an
index buffer, vertex buffer or render target as those can also be handy
to remember which was which.

mark_snapshot is still respected for backwards compatibility, but throws
a deprecation warning if hunting is enabled.