Copy on mark: Check for existing shaders of any type
When copy on mark is activated we check if an existing HLSL shader
exists and dump it if not. Since we may have an assembly shader in
ShaderFixes this may result in us dumping out a conflicting file. While
we already have code that will warn that this has happened, it does not
prevent it from happening.

This patch adds an initial check for any existing shaders of any type
before going through with the copy on mark procedure. This is outside of
the existing code flow as it is intended to also support dumping
assembly shaders in a subsequent patch. This detection also checks for
the existence of cached .bin files if no .txt file exists and warns
about those as well.

This currently does not touch the files if they already exist - that
will be added back in the following patch.