Touch existing files when marked
This restores the use case of touching an existing shader when it is
marked to make it easy to find in explorer. We used to do this by
appending a space to the file, but that resulted in spurious diff hunks.
Instead, we now simply update the timestamp on the file directly, using
the same code as for shader cache management.

This will not touch .bin files found in ShaderFixes, as that could cause
a stale .bin file to be marked as valid.

The old code to detect and touch existing HLSL files has been removed in
favour of the new procedure.