Touch existing shader on mark: Update ShaderFixes timestamp as well
Explorer (at least on Windows 10) does seem to notice when we touch the
timestamp on the existing shader as the modified time column gets
updated, but it does not seem to automatically re-sort the directory.

This patch also touches the ShaderFixes folder* whenever we touch an
existing shader inside it, which causes explorer to immediately refresh
it's directory listing.

* which requires the FILE_FLAG_BACKUP_SEMANTICS flag to CreateFile...
nice of them to use a function with a relevant name and a flag named
after what is is for /s