Refactor CopyToFixes and WriteHLSL
This moves HLSL specific code from CopyToFixes into WriteHLSL to make
CopyToFixes easier to follow and allow it to be used for assembly as
well later.

This also switches a few large C++ string copies to pass by pointer in
these routines, and the WriteHLSL routine was relocated below the
functions it now refers to.