"marking_actions=hlsl asm" try HLSL and fallback to Asm on failure
This is intended to be the default mode we ship in. It will first try
decompiling a shader to HLSL and reloading it, and if that fails will
remove the freshly decompiled shader (but not existing shaders that
happen to have errors) and automatically switch to assembly.

If the shaderhacker would rather correct the error in HLSL and work with
that, they can choose marking_actions=hlsl instead.

Obviously it cannot detect cases where the shader successfully reloads
but still has errors causing visual defects, but the idea is if those
are common in a given game the shaderhacker could choose to switch to
marking_actions=asm.