HLSL Decompiler: Remove inappropriate compiler warning suppression
This is not a benign warning where these values are passed on the stack
as it can lead to scanf reading the wrong number of bytes, and this
warning should never have been suppressed.

These bugs have previously been fixed by:
    "Fix various compiler warnings when building with VS2015"
    commit 6673f220150390c9a5e91e62d4be3eb3924ac7c4
    Author: Ian Munsie <darkstarsword@gmail.com>
    Date:   Sat Jan 23 20:05:45 2016 +1100

As such, this suppression no longer serves any purpose but may hide
future bugs, so remove it.