HLSL Decompiler: Add a macro to simplify passing sizeof(buf) to scanf
I'm about to add more parsing routines to the decompiler and am not
really happy with the cast to int boilerplate necessary for each buffer
size passed to scanf_s. This adds a simple macro to add UCOUNTOF as an
alias around _countof to reduce the amount of boilerplate.

We use _ucountof() instead of sizeof() as it checks to make sure it is
an array for extra safety.