HLSL Decompiler: First pass support for parsing structured buffers
This adds the first piece necessary to support structured buffers in the
decompiler. It parses the "Resource bind info" from the assembly
comments and pulls out the HLSL structure it contains and matches the
structure type name when a structure is bound to a register.

It supports both SM4 and SM5 structured buffers - notably SM4 lacks a
type name, so for that case it creates one.

Test case .chk files are updated to keep an eye on how the shaders
change.