Add temporary HLSL test harness and StructuredBuffer test case
This adds a Unix shell script to serve as a temporary test harness for
the decompiler and [dis]assembler to facilitate regression testing and
test driven development.

The idea is that we will eventually include the test cases as a standard
part of our build - however, that will require switching to something
that can be natively supported in MSVC (whether this is a windows batch
file, powershell script or adopting an existing test harness), and for
now I preferred to go with something trivially simple and that I already
am familiar with to get started.

This adds several test cases for StructuredBuffer support in the HLSL
decompiler. This follows the Test Driven Development model, where these
tests are written prior to support being added - hence, they currently
fail. Once support is added these tests should begin to pass and provide
a measurement of success.

The resource_types test case for the assembler already included test
cases for StructuredBuffers, but it also included a number of other
types that the decompiler lacks support for, so cannot not be used as a
test for this single feature.